Feature: expose price records via /api/prices endpoint

+++ src/main/java/acessa/PriceComparator/PriceCsvRow.java
@@ -10,6 +10,6 @@ public class PriceCsvRow {
 	@CsvBindByName private String brand;
 	@CsvBindByName private double package_quantity;
 	@CsvBindByName private String package_unit;
-	@CsvBindByName private double price;
+	@CsvBindByName private float price;
 	@CsvBindByName private String currency;
 }
